cleanup: improve error handling
